Refactor using AppSettings class, make capability classes only internal accessible for DeviceTransferMethodManager
