+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/method/AppSettings.kt
@@ -1,0 +1,5 @@
+package at.asitplus.wallet.app.common.iso.transfer.method
+
+internal expect class AppSettings() {
+    fun open(platformContext: PlatformContext)
+}
