+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/method/BluetoothInfo.kt
@@ -2,7 +2,7 @@ package at.asitplus.wallet.app.common.iso.transfer.method
 
 import androidx.compose.runtime.Composable
 
-expect class BluetoothInfo() {
+internal expect class BluetoothInfo() {
     @Composable
     fun isBluetoothEnabled(): Boolean
     fun openBluetoothSettings(platformContext: PlatformContext)
