+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/method/DeviceTransferMethodManager.kt
@@ -7,12 +7,16 @@ import androidx.compose.runtime.Stable
 class DeviceTransferMethodManager {
     private val bluetoothInfo = BluetoothInfo()
     private val nfcInfo = NfcInfo()
+    private val appSettings = AppSettings()
 
     @Composable
     fun isBluetoothEnabled(): Boolean = bluetoothInfo.isBluetoothEnabled()
-    fun goToBluetoothSettings(platformContext: PlatformContext) = bluetoothInfo.openBluetoothSettings(platformContext)
+    fun goToBluetoothSettings(platformContext: PlatformContext) =
+        bluetoothInfo.openBluetoothSettings(platformContext)
 
     @Composable
     fun isNfcEnabled(): Boolean = nfcInfo.isNfcEnabled()
     fun goToNfcSettings(platformContext: PlatformContext) = nfcInfo.openNfcSettings(platformContext)
+
+    fun openAppSettings(platformContext: PlatformContext) = appSettings.open(platformContext)
 }
