+++ shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/method/NfcInfo.kt
@@ -2,7 +2,7 @@ package at.asitplus.wallet.app.common.iso.transfer.method
 
 import androidx.compose.runtime.Composable
 
-expect class NfcInfo() {
+internal expect class NfcInfo() {
     @Composable
     fun isNfcEnabled(): Boolean
     fun openNfcSettings(platformContext: PlatformContext)
