--- shared/src/commonMain/kotlin/at/asitplus/wallet/app/common/iso/transfer/method/openAppSettings.kt
@@ -1,6 +1,0 @@
-package at.asitplus.wallet.app.common.iso.transfer.method
-
-/**
- * Opens app settings page so the user can manage permissions.
- */
-expect fun openAppSettings(platformContext: PlatformContext)
