+++ shared/src/commonMain/kotlin/ui/views/iso/common/MissingPreconditionView.kt
@@ -24,7 +24,6 @@ import at.asitplus.valera.resources.Res
 import at.asitplus.valera.resources.heading_label_missing_precondition
 import at.asitplus.wallet.app.common.iso.transfer.method.DeviceTransferMethodManager
 import at.asitplus.wallet.app.common.iso.transfer.method.PlatformContext
-import at.asitplus.wallet.app.common.iso.transfer.method.openAppSettings
 import at.asitplus.wallet.app.common.iso.transfer.state.PreconditionState
 import at.asitplus.wallet.app.common.iso.transfer.state.TransferSettingsState
 import org.jetbrains.compose.resources.stringResource
@@ -118,7 +117,9 @@ fun MissingPreconditionViewBody(
                 blePermissionState.launchPermissionRequest()
             }
             MissingBluetoothPermissionView(
-                onOpenAppPermissionSettings = { openAppSettings(platformContext) }
+                onOpenAppPermissionSettings = {
+                    deviceTransferMethodManager.openAppSettings(platformContext)
+                }
             )
         }
     }
